Drag and drop letters

+++ engine/source/platform/types.posix.h
@@ -29,8 +29,7 @@
 // size_t is needed to overload new
 // size_t tends to be OS and compiler specific and may need to 
 // be if/def'ed in the future
-#include <cstddef>
-typedef size_t  dsize_t;      
+typedef unsigned int  dsize_t;      
 
 
 /** Platform dependent file date-time structure.  The defination of this structure
